# api/openapi.yaml
openapi: 3.1.0
info:
  title: frp-panel API
  version: 2.0.0-edge
  license:
    name: AGPL-3.0
    identifier: AGPL-3.0-only
servers:
  - url: /api/v2
paths:
  /health:
    get:
      operationId: getHealth
      security: []
      responses:
        '200':
          description: Master is ready.
          content:
            application/json:
              schema:
                type: object
                required: [status, time, protocol]
                properties:
                  status: { const: ok }
                  time: { type: string, format: date-time }
                  protocol: { const: '2' }
  /bootstrap-status:
    get:
      operationId: getBootstrapStatus
      security: []
      responses:
        '200':
          description: Minimal first-run state used to select setup or login UI.
          content:
            application/json:
              schema:
                type: object
                required: [registrationEnabled, ownerExists, canCreateOwner]
                properties:
                  registrationEnabled: { type: boolean }
                  ownerExists: { type: boolean }
                  canCreateOwner: { type: boolean }
  /capabilities:
    get:
      operationId: getCapabilities
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Master and Client Agent capabilities.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /enrollments:
    post:
      operationId: createAgentEnrollment
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clientId]
              properties:
                clientId: { type: string, pattern: '^[A-Za-z0-9_-]+$' }
      responses:
        '201':
          description: A one-use enrollment valid for ten minutes.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Enrollment' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /agent/enroll:
    post:
      operationId: redeemAgentEnrollment
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string, minLength: 32, maxLength: 256 }
      responses:
        '200':
          description: Permanent credential returned once and never stored in plaintext.
          headers:
            Cache-Control: { schema: { const: no-store } }
          content:
            application/json:
              schema:
                type: object
                required: [clientId, secret]
                properties:
                  clientId: { type: string }
                  secret: { type: string, writeOnly: true }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
  /server-enrollments:
    post:
      operationId: createServerEnrollment
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [serverId, serverIp]
              properties:
                serverId: { type: string, pattern: '^[A-Za-z0-9_-]+$' }
                serverIp: { type: string, description: Public IP address or DNS hostname without a scheme or port. }
                bindPort: { type: integer, minimum: 1024, maximum: 65535, default: 7000 }
      responses:
        '201':
          description: One-use FRPS enrollment valid for ten minutes.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ServerEnrollment' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /server/enroll:
    post:
      operationId: redeemServerEnrollment
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string, minLength: 32, maxLength: 256 }
      responses:
        '200':
          description: Permanent FRPS credential returned once and never stored in plaintext by Master.
          headers:
            Cache-Control: { schema: { const: no-store } }
          content:
            application/json:
              schema:
                type: object
                required: [serverId, secret]
                properties:
                  serverId: { type: string }
                  secret: { type: string, writeOnly: true }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
  /account:
    get:
      operationId: getAccount
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: The signed-in account profile.
          content:
            application/json:
              schema:
                type: object
                required: [username, email, role]
                properties:
                  username: { type: string }
                  email: { type: string, format: email }
                  role: { type: string }
        '401': { $ref: '#/components/responses/Problem' }
  /account/password:
    post:
      operationId: changePassword
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string, format: password }
                newPassword: { type: string, format: password, minLength: 12 }
      responses:
        '200':
          description: Password changed and the current session invalidated.
          content:
            application/json:
              schema:
                type: object
                required: [reauthenticate]
                properties:
                  reauthenticate: { const: true }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
  /tunnels:
    get:
      operationId: listTunnels
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Managed TCP and UDP tunnels.
          content:
            application/json:
              schema:
                type: object
                required: [tunnels]
                properties:
                  tunnels:
                    type: array
                    items: { $ref: '#/components/schemas/Tunnel' }
        '401': { $ref: '#/components/responses/Problem' }
    post:
      operationId: createTunnel
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TunnelMutation' }
      responses:
        '201':
          description: Tunnel saved; its Client-to-Server FRPC connection is created or reused automatically.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tunnel' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
    delete:
      operationId: deleteTunnel
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, clientId, serverId]
              properties:
                name: { type: string }
                clientId: { type: string }
                serverId: { type: string }
      responses:
        '204': { description: Tunnel removed. }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: frp-panel-cookie
  schemas:
    Enrollment:
      type: object
      required: [clientId, token, expiresAt]
      properties:
        clientId: { type: string }
        token: { type: string, writeOnly: true }
        expiresAt: { type: string, format: date-time }
    ServerEnrollment:
      type: object
      required: [serverId, token, expiresAt]
      properties:
        serverId: { type: string }
        token: { type: string, writeOnly: true }
        expiresAt: { type: string, format: date-time }
    Tunnel:
      type: object
      required: [name, clientId, serverId, type, localHost, localPort, remotePort, stopped]
      properties:
        name: { type: string }
        clientId: { type: string }
        serverId: { type: string }
        type: { type: string, enum: [tcp, udp] }
        localHost: { type: string }
        localPort: { type: integer, minimum: 1, maximum: 65535 }
        remotePort: { type: integer, minimum: 1024, maximum: 65535 }
        stopped: { type: boolean }
    TunnelMutation:
      type: object
      required: [name, clientId, serverId, type, localPort, remotePort]
      properties:
        name: { type: string, pattern: '^[A-Za-z0-9_-]+$' }
        clientId: { type: string }
        serverId: { type: string }
        type: { type: string, enum: [tcp, udp] }
        localHost: { type: string, default: 127.0.0.1 }
        localPort: { type: integer, minimum: 1, maximum: 65535 }
        remotePort: { type: integer, minimum: 1024, maximum: 65535 }
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
  responses:
    Problem:
      description: RFC 9457 problem detail.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }